requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Django==2.0.10
django-cors-headers==2.4.0
django-extensions==2.1.4
django-filter==2.0.0
django-phonenumber-field==2.1.0
django-twilio==0.9.3
django-user-agents-last==0.3.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.8
Pillow==5.4.0
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.7
PyYAML==5.1
requests==2.21.0
six==1.12.0
South==1.0.2
twilio==6.23.1
typing==3.6.6
ua-parser==0.8.0
urllib3==1.24.1
user-agents==1.1.0
